Improving logging. The now separate thread Indexer was not logging anymore
as per tomcat documentation :
«
Old applications that still use System.out or System.err can be tricked by setting swallowOutput attribute on a Context. If the attribute is set to true, the calls to System.out/err during request processing will be intercepted, and their output will be fed to the logging subsystem using the javax.servlet.ServletContext.log(...) calls.

Note, that the swallowOutput feature is actually a trick, and it has its limitations. It works only with direct calls to System.out/err, and only during request processing cycle. It may not work in other threads that might be created by the application.
»

reference : https://tomcat.apache.org/tomcat-8.5-doc/logging.html